bridge-autopilot-wrapper/requirements.txt
PyQt5
python-dotenv
pyzmq